move get_processed_sibling_instruction and get_stack_height back to solana-program to avoid breaking change
