+++ sdk/program/src/instruction.rs
@@ -1,5 +1,71 @@
 pub use solana_instruction::{
-    checked_add, error::InstructionError, get_processed_sibling_instruction, get_stack_height,
-    AccountMeta, CompiledInstruction, Instruction, ProcessedSiblingInstruction,
-    TRANSACTION_LEVEL_STACK_HEIGHT,
+    checked_add, error::InstructionError, AccountMeta, CompiledInstruction, Instruction,
+    ProcessedSiblingInstruction, TRANSACTION_LEVEL_STACK_HEIGHT,
 };
+
+/// Returns a sibling instruction from the processed sibling instruction list.
+///
+/// The processed sibling instruction list is a reverse-ordered list of
+/// successfully processed sibling instructions. For example, given the call flow:
+///
+/// A
+/// B -> C -> D
+/// B -> E
+/// B -> F
+///
+/// Then B's processed sibling instruction list is: `[A]`
+/// Then F's processed sibling instruction list is: `[E, C]`
+pub fn get_processed_sibling_instruction(index: usize) -> Option<Instruction> {
+    #[cfg(target_os = "solana")]
+    {
+        let mut meta = ProcessedSiblingInstruction::default();
+        let mut program_id = solana_pubkey::Pubkey::default();
+
+        if 1 == unsafe {
+            solana_instruction::syscalls::sol_get_processed_sibling_instruction(
+                index as u64,
+                &mut meta,
+                &mut program_id,
+                &mut u8::default(),
+                &mut AccountMeta::default(),
+            )
+        } {
+            let mut data = Vec::new();
+            let mut accounts = Vec::new();
+            data.resize_with(meta.data_len as usize, u8::default);
+            accounts.resize_with(meta.accounts_len as usize, AccountMeta::default);
+
+            let _ = unsafe {
+                sol_get_processed_sibling_instruction(
+                    index as u64,
+                    &mut meta,
+                    &mut program_id,
+                    data.as_mut_ptr(),
+                    accounts.as_mut_ptr(),
+                )
+            };
+
+            Some(Instruction::new_with_bytes(program_id, &data, accounts))
+        } else {
+            None
+        }
+    }
+
+    #[cfg(not(target_os = "solana"))]
+    crate::program_stubs::sol_get_processed_sibling_instruction(index)
+}
+
+/// Get the current stack height, transaction-level instructions are height
+/// TRANSACTION_LEVEL_STACK_HEIGHT, fist invoked inner instruction is height
+/// TRANSACTION_LEVEL_STACK_HEIGHT + 1, etc...
+pub fn get_stack_height() -> usize {
+    #[cfg(target_os = "solana")]
+    unsafe {
+        solana_instruction::syscalls::sol_get_stack_height() as usize
+    }
+
+    #[cfg(not(target_os = "solana"))]
+    {
+        crate::program_stubs::sol_get_stack_height() as usize
+    }
+}
